Port d3 code into separate file
